# Improve JAccess::getAssetRules() Performance ~50%