@sovainfo 2015-04-06
Confirmed.<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="http://issues.joomla.org/tracker/joomla-cms/6681">issues.joomla.org/joomla-cms/6681</a>.</sub>