@AnneKlapwijk 2016-04-15
I have tested this item :red_circle: unsuccessfully on cb9f34fd4116695479474c5e7f2da06ddb8af0a4<br /><br />Tested in /administrator/index.php?option=com_content&view=articles using the Joomla!Debug Console with 68 articles (default testing sample content).
Didn't find any new bugs caused by this patch.

**Before patch**
49 Queries Logged 292.32 ms
49 Queries Logged 93.25 ms
49 Queries Logged 50.97 ms
49 Queries Logged 97.60 ms
49 Queries Logged 68.06 ms
49 Queries Logged 92.23 ms

**After patch**
49 Queries Logged 205.16 ms
49 Queries Logged 108.98 ms
49 Queries Logged 96.91 ms
49 Queries Logged 83.46 ms
49 Queries Logged 131.34 ms
49 Queries Logged 93.73 ms

**Stats**
 OS Windows
 PHP 5.6.11
 MySQLi 5.6.25
 Time 12:37
 Caching Disabled
 GZip Disabled
 Articles 68<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6681">issues.joomla.org/joomla-cms/6681</a>.</sub>