@superknutsel 2016-04-15
I've tested unsuccessful using the Debug console.
No issues where found applying the patch.

before patch 72 articles
1	15.61ms
2	18.00ms
3	18.85ms
4	20.31ms
5	36.46ms

after patch 72 articles
1	18,26
2	18,99
3	18,98
4	19,28
5	20,20

PHP Version   7.0.4
DB Type       mysqli
DB Version    5.5.5
CMS Version   3.5.2<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6681">issues.joomla.org/joomla-cms/6681</a>.</sub>