@roland-d 2016-06-24
Hello @orware 

Thank you for your contribution.

The last comment here was on April 15th. We have had 2 unsuccessful tests, Is this issue/pull request still valid?
If no reply is received within 4 weeks we will close this issue.

Thanks for understanding! <hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6681">issues.joomla.org/joomla-cms/6681</a>.</sub>