@brianteeman 2016-07-21
I am closing this PR at this time for the reasons previously stated. It can always be re-opened if required<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6681">issues.joomla.org/joomla-cms/6681</a>.</sub>